requirements.txt
python-dotenv==1.2.1
google-genai==1.56.0
google-adk==1.21.0
requests~=2.32.5